Separate the extensions for CoreGraphics from Image.swift
